Upgrade Android template as per https://github.com/facebook/react-native/commit/49f20f41546e3ba8e7fe43c84c4c701684d0434d, following instructions 0.29 release notes at https://github.com/facebook/react-native/releases/tag/v0.29.0
